requirements.txt
django==4.2.4
django-debug-toolbar==4.2.0
python-dotenv==1.0.0
djangorestframework==3.14.0
mysqlclient==